Move `BackPressure`, `Simulate`, `Drivable` and `Test`
To prevent cyclic dependencies when implementing TH functions that
generate tuple instances for these typeclasses.
