+++ clash-protocols/clash-protocols.cabal
@@ -145,11 +145,11 @@ library
     Protocols.Axi4.WriteData
     Protocols.Axi4.WriteResponse
     Protocols.Df
-    Protocols.Internal
     Protocols.DfConv
     Protocols.Hedgehog
     Protocols.Hedgehog.Internal
     Protocols.Idle
+    Protocols.Internal
     Protocols.Internal.TH
     Protocols.Wishbone
     Protocols.Wishbone.Standard
@@ -165,8 +165,9 @@ library
   autogen-modules:    Paths_clash_protocols
 
   other-modules:
-    Protocols.Internal.Types
     Paths_clash_protocols
+    Protocols.Hedgehog.Types
+    Protocols.Internal.Types
 
   default-language: Haskell2010
